Fix stats estimate for UDFs.
It's going to be a bit hard to get estimates for UDFs. Currently they
return -1. We had a place where we would multiply by the selectivity
without checking it is non-negative, which failed a check in the planner.

It's unclear what the best option is. We could propagate -1 up the entire
plan tree but that's going to lead to some really whacky plans. This patch
ignores unknown exprs (gives them selectivity 1) which can be interpreted
as being most pessimistic.

Change-Id: Ib54a4848c5152d7475c9ddf82cdcb8aa4de58781
Reviewed-on: http://gerrit.ent.cloudera.com:8080/738
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Nong Li <nong@cloudera.com>