Update mem pool to support TryAllocate and RcFileScanner to use it.
We've got users exceeding the mem limit by a large amount (e.g. 20GB). This
is likely because we allow the limit to be exceeded and check after. This
usually works pretty well except for some allocations that we know can be
very large. The RcFile scanner has one such place and we can extend using
TryAllocate() to more places in the future.

Change-Id: Idd62b9be621be05008a8b6ea812e9763504e3527
Reviewed-on: http://gerrit.ent.cloudera.com:8080/758
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: Nong Li <nong@cloudera.com>