Allow different Kerberos principals for front- and back- end services.
This patch allows different principals for client and server
services. The server principal is specified by --be_principal (unless
not set, in which case --principal is used). Kinit is run for the server
principal only.

This patch also cleans up a few errors in the way Kinit was run, so that
failure can be detected more accurately.

Change-Id: I30bf747666caca17772ad30a7d14eab03ce799d6
Reviewed-on: http://gerrit.ent.cloudera.com:8080/890
Reviewed-by: Alan Choi <alan@cloudera.com>
Tested-by: jenkins