Remove deadlock if GetResultSetMetadata() is called on closed session
This patch fixes a bug where, if withSession() returned an
error, GetResultSetMetadata() could return without releasing a query
exec state's lock.

The exact order of events required to trigger the bug is unlikely, but
not unimaginable:

* GetResultsMetadata() retrieves a query exec state and takes its lock.
* The associated session expires or closes for some
  reason (i.e. disconnection)
* GetResultsMetadata() calls withSession() on that session, realises
  the session is no longer valid and returns without releasing the exec
  state lock.

While possible, this sequence is unlikely if session-expiry is *not*
enabled, because the session would have to have been closed either
explicitly (via CloseSession()) or implicitly (via disconnect) on a
separate connection to the one that GetResultsMetadata() is sent on.

Change-Id: I4b7481b1c87a25a7801d8a5736bfc0c9980384cf
Reviewed-on: http://gerrit.ent.cloudera.com:8080/999
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins