Fixup error message returned to user for tables with incomplete metadata
When there was an error loading the table metadata the Catalog Server
includes (among other things) the table loading error message(s) and
the call stack where the error happened. The error messages returned
to the user previously included the call stack, but with this change
the call remote (catalog service) call stack is only output to the impalad
log when the table is accessed by the user. This is done by annotating
and appending the remote call stack to the local call stack.

Change-Id: Icf46a3d7fc15d8b0a8e59564722fdb991b074618
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1063
Tested-by: jenkins
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>