IMPALA-142: Expose memory usage metrics from the JVM
This patch adds metrics to track the memory usage of the various memory
pools inside the JVM. The metrics are fetched on demand when their value
is calculated; therefore they are not free to query but doing so is not
on a critical path.

No GC statistics are captured yet, but extending this patch is the
natural way to do so.

Change-Id: Ibb80ef8b27987295a8f8fc24c8b2995a58c299dc
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1074
Tested-by: jenkins
Reviewed-by: Henry Robinson <henry@cloudera.com>