Properly set TAccessLevel when loading partitions from thrift to fix READ_ONLY error msg
When loading partitions from thrift we were not setting the accessLevel. This caused
the partition to inherit the accessLevel from the parent table. This didn't have any
functional impact, but caused the READ_ONLY error messages on INSERT to fail to show the
proper HDFS paths.
Note: There were already analyzer tests to validate the error messages, but those didn't
catch this because they don't exercise the end-to-end loading to/from thrift code.

Change-Id: I5c82458d59a638ed8e771122830116419f78dfc4
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1083
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: jenkins