Update tests to remove uneeded calls to invalidate all catalog metadata
All tables/databases/functions are dropped via Impala so there is no
need to invalidate all catalog metadata. Not only is this operation
expensive, it can cause other tests to fail when executed in parallel
since it has the potential to temporarily reset some direct DDL operations
that have been applied to the local catalog.

Change-Id: Ia9a48d8a0dbe2eb5928b41950d17a35e8991fe22
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1090
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: jenkins