IMPALA-640: ExecNodes should close their children as early as possible
We can close some nodes before the entire fragment is closed in order to
release resources that are no longer needed. However, we must be careful
and ensure that we do not close nodes if they own memory that is still
referenced. In order to ensure we acquire all resources that we need, we
will only call Close() on nodes for which we have consumed all of the
row batches and acquired their resources. There are a number of cases
where this is OK: blocking nodes that consume all input in Open(), the
right child of a BlockingJoinNode (which is fully consumed in Open()),
children of a MergeNode that are fully consumed (but not the child that
is still open if a LIMIT is reached).
Because some nodes will be closed early, the Close() method has been made
idempotent.

Change-Id: I4424fe07ab874342fb20390c70328c31238d27ed
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1020
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins