Fix Parquet compatibilty bug.
We were using the number of rows in the file to determine when a
column reader was finished, but this only works for files with a
single row group. This means we could read files produced by Impala
but necessarily those produced by MR. Instead, we now use the number
of values in the column directly.

Change-Id: I6ae38d11d71efc82ce4d140c5be95f880c16e092
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1179
Reviewed-by: Ishaan Joshi <ishaan@cloudera.com>
Tested-by: jenkins