[CDH5] Upgrade to latest MiniLlama (0fde2e0)
Note: This uses only the MiniLlama tarball from the Llama repository,
which provides the minimal libraries we need for development. Before, we
somehow amalgamated the Llama and MiniLlama tarballs.

To work around a 'feature' introduced upstream in YARN, I've manually
added a fairscheduler.xml file to $LLAMA_HOME/conf with three queues.

Change-Id: I69e36ed9da9adeb76afb9801d9031b02d8d6232c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1278
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Henry Robinson <henry@cloudera.com>