[CDH4] CDH-17085: Remove import of Parquet Hive SerDe class that no longer exists
This change is to unblock Impala on CDH5. The Hive Parquet SerDe has changed Java
packages from:

"parquet.hive.serde.ParquetHiveSerDe"
to
"org.apache.hadoop.hive.ql.io.parquet.serde.ParquetHiveSerDe"

Because of how our build/packaging is configured, this doesn't cause a compile
time error, because the old class is available at build time, but causes a runtime
failure because the class doesn't exist in an actual CDH5 installation. The fix
is to just use a string as the class name, instead of doing an import.

There is some additional work that needs to happen to verify compatibility
of the new/old SerDe and add support for any new INPUT/OUTPUT format classes.
That is tracked by CDH-17085.

Note: This commit also exists in CDH5 (a57497), but later decided to backport to CDH4.

Change-Id: I9cd0a2d8c2aae03d7bbb9d53d5b34747af69cd12
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1398
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: Lenni Kuff <lskuff@cloudera.com>
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1798
Tested-by: jenkins