IMP-1133: Add pause before renewing Kerberos ticket for the first time.
See https://issues.cloudera.org/browse/hue-640 for a full description of the issue. The
summary is that kinit -R can't come less than a second after the initial kinit.

In order to separate out two separate calls this patch adds a method to run a shell
command and gather the output.

Change-Id: I65b79150c1df0d3f08f4200f59088eab644b0f35
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1387
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: jenkins